+++ slsDetectorSoftware/gotthardDetectorServer/firmware_funcs.c
@@ -1208,8 +1208,8 @@ int setDACRegister(int idac, int val, int imod) {
 }
 
 
-float getTemperature(int tempSensor, int imod){
-  float val;
+double getTemperature(int tempSensor, int imod){
+  double val;
   char cTempSensor[2][100]={"ADCs/ASICs","VRs/FPGAs"}; 
   imod=0;//ignoring more than 1 mod for now
   int i,j,repeats=6;
@@ -1236,7 +1236,7 @@ float getTemperature(int tempSensor, int imod){
   }
 
   bus_w(TEMP_IN_REG,(T1_CLK_BIT)|(T1_CS_BIT)|(T2_CLK_BIT)|(T2_CS_BIT));//standby
-  val=((float)tempVal)/4.0;
+  val=((double)tempVal)/4.0;
 
 #ifdef VERBOSE
    printf("Temperature of module:%d for the %s is %.2fC\n",imod,cTempSensor[tempSensor],val);
