+++ slsDetectorSoftware/gotthardDetectorServer/server_funcs.c
@@ -899,13 +899,13 @@ int read_register(int file_des) {
 
 int set_dac(int file_des) {
 	//default:all mods
-	float retval;
+	double retval;
 	int ret=OK;
 	int arg[2];
 	enum dacIndex ind;
 	int imod;
 	int n;
-	float val;
+	double val;
 	int idac=0;
 
 	sprintf(mess,"Can't set DAC\n");
@@ -1028,7 +1028,7 @@ int set_dac(int file_des) {
 
 int get_adc(int file_des) {
 	//default: mod 0
-	float retval;
+	double retval;
 	int ret=OK;
 	int arg[2];
 	enum dacIndex ind;
@@ -1373,8 +1373,8 @@ int set_module(int file_des) {
   sls_detector_module myModule;
   int *myChip=malloc(NCHIP*sizeof(int));
   int *myChan=malloc(NCHIP*NCHAN*sizeof(int));
-  float *myDac=malloc(NDAC*sizeof(int));
-  float *myAdc=malloc(NADC*sizeof(int));
+  double *myDac=malloc(NDAC*sizeof(int));
+  double *myAdc=malloc(NADC*sizeof(int));
   int retval, n;
   int ret=OK;
   int dr;// ow;
@@ -1492,8 +1492,8 @@ int get_module(int file_des) {
   sls_detector_module myModule;
   int *myChip=malloc(NCHIP*sizeof(int));
   int *myChan=malloc(NCHIP*NCHAN*sizeof(int));
-  float *myDac=malloc(NDAC*sizeof(int));
-  float *myAdc=malloc(NADC*sizeof(int));
+  double *myDac=malloc(NDAC*sizeof(int));
+  double *myAdc=malloc(NADC*sizeof(int));
 
 
   if (myDac)
