+++ slsDetectorSoftware/gotthardDetectorServer/trimming_funcs.c
@@ -161,11 +161,11 @@ int choose_vthresh_and_vtrim(int countlim,  int nsigma, int im) {
 
   int *fifodata;
 
-  float vthreshmean, vthreshSTDev;
+  double vthreshmean, vthreshSTDev;
   int *thrmi, *thrma;
-  float c;
-  float b=BVTRIM;       
-  float a=AVTRIM; 
+  double c;
+  double b=BVTRIM;
+  double a=AVTRIM;
   int *trim;
   int ich, imod, ichan;
   int nvalid=0;
@@ -439,7 +439,7 @@ int trim_with_level(int countlim, int im) {
 int kth_smallest(int *a, int n, int k)
 {
     register int i,j,l,m ;
-    register float x ;
+    register double x ;
 
     l=0 ; m=n-1 ;
     while (l<m) {
@@ -466,7 +466,7 @@ int ave(int *a, int n)
 {
   int av=0,i;
   for (i=0; i<n; i++)
-    av=av+((float)*(a+i))/((float)n);
+    av=av+((double)*(a+i))/((double)n);
   return av;
 }
